+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added /rfuemojis command to list all available emojis along with their shortest alias
 - Added option to configure custom rare sea creature party messages per creature in /rfuscedit
 - Added 1 new achievements!
+- Added Lilypad Size Display feature
 
 ### Changes
 - Removed 1.21.11 Support